Added a git template.
The pre-push hook prevents submodule users from pushing to a remote
if any of its submodules are not themselves pushed to origin.

linkdots.sh was updated to copy the template directory to ~/.gittemplate
and then call git config --global init.templatedir=~/.gittemplate. From
then on, any git repo in which `git init` is run will get the hook
installed. Note `git init` can safely be called in an existing repo to
get the hook.